Refactor - moved forwarding functions to its own library
