+++ src/lib/orionld/common/orionldState.cpp
@@ -102,6 +102,8 @@ OrionldPhase      orionldPhase             = OrionldPhaseStartup;
 bool              orionldStartup           = true;
 char              pgPortString[16];
 char              mongoServerVersion[32];
+char              userAgentHeaderNoLF[64];     // "User-Agent: orionld/" + ORIONLD_VERSION - initialized in orionldServiceInit()
+char              hostHeader[256];             // Host: xxx
 
 
 //
